[#3]feat: make ItemList file and move ItemList type to it
